Feat: Header and Footer template.
The Header template will return any `script` and `link[ref=stylesheet]` tags
that appear above the `header` tag.  This is in addition to the header tag and
its children.

The Footer template will return any `script` tags that appear below the `footer`
tag.  This is in addition to the `footer` tag and its children.