Feature: netcall integration (#7)
* Feat: Template with empty main tag.

The presence of the `nocontent` HTTP query parameter in requests for the
ModernGov template will produce a template with an empty main tag.

* CI: PHP 8.2 for test runs under Drupal 10.x.

* Feat: Header and Footer template.

The `header` and `footer` HTTP query parameters will return only the `header`
and `footer` tags and their children from the ModernGov template.

* Feat: Header and Footer template.

The Header template will return any `script` and `link[ref=stylesheet]` tags
that appear above the `header` tag.  This is in addition to the header tag and
its children.

The Footer template will return any `script` tags that appear below the `footer`
tag.  This is in addition to the `footer` tag and its children.

* Chore: Coding standards fixes.

* CI/CD: Github actions test runner.

Copying changes from upstream text runner file.

* Fix: Deactivating BigPipe to suppress PHP warnings.

BigPipe functionality is not necessary as we are mostly serving these templates
to anonymous users.  This can be reviewed later if necessary.